% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{projr_cat_changelog}
\alias{projr_cat_changelog}
\title{Cat useful information}
\usage{
projr_cat_changelog(n_row = 10)
}
\arguments{
\item{n_row}{integer.
Number of rows to cat.}
}
\description{
Cat useful information.
.cat_changelog`:
}
