% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-pre.R
\name{.build_debug_git_info}
\alias{.build_debug_git_info}
\title{Output Git information for debug}
\usage{
.build_debug_git_info(output_level = "std")
}
\arguments{
\item{output_level}{Character. Current output level}
}
\description{
Output Git information for debug
}
\keyword{internal}
