% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-output.R
\name{.cli_debug}
\alias{.cli_debug}
\title{Show a debug message (only shown at debug level)}
\usage{
.cli_debug(..., output_level = "std", .envir = parent.frame(), log_file = NULL)
}
\arguments{
\item{...}{Message components passed to cli::cli_text}

\item{output_level}{Character. Current output level.}

\item{.envir}{Environment for variable evaluation}

\item{log_file}{Character. Path to log file (optional).}
}
\description{
Show a debug message (only shown at debug level)
}
\keyword{internal}
