% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-output.R
\name{.cli_output_level_get}
\alias{.cli_output_level_get}
\title{Get the current output level}
\usage{
.cli_output_level_get()
}
\value{
Character. The output level to use.
}
\description{
Get the current output level
}
\keyword{internal}
