% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-output.R
\name{.cli_process_start}
\alias{.cli_process_start}
\title{Start a process with a spinner/status indicator}
\usage{
.cli_process_start(...)
}
\arguments{
\item{...}{Message components}
}
\value{
Process ID for cli_process_done
}
\description{
Start a process with a spinner/status indicator
}
\keyword{internal}
