% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-output.R
\name{.cli_stage_header}
\alias{.cli_stage_header}
\title{Show a stage header}
\usage{
.cli_stage_header(stage_name, build_type = "output")
}
\arguments{
\item{stage_name}{Character. Name of the build stage.}

\item{build_type}{Character. Type of build ("dev" or "output").}
}
\description{
Show a stage header
}
\keyword{internal}
