% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-output.R
\name{.cli_step}
\alias{.cli_step}
\title{Show a step message (sub-item in a stage)}
\usage{
.cli_step(..., .envir = parent.frame())
}
\arguments{
\item{...}{Message components passed to cli::cli_li}

\item{.envir}{Environment for variable evaluation}
}
\description{
Show a step message (sub-item in a stage)
}
\keyword{internal}
