% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-output.R
\name{.cli_success}
\alias{.cli_success}
\title{Show a success message}
\usage{
.cli_success(..., .envir = parent.frame())
}
\arguments{
\item{...}{Message components passed to cli::cli_alert_success}

\item{.envir}{Environment for variable evaluation}
}
\description{
Show a success message
}
\keyword{internal}
