% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.log_build_append}
\alias{.log_build_append}
\title{Append message to build log}
\usage{
.log_build_append(message, level = "info")
}
\arguments{
\item{message}{Character. Message to append.}

\item{level}{Character. Message level (info, debug, success, etc.)}
}
\description{
Append message to build log
}
\keyword{internal}
