% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.log_build_section}
\alias{.log_build_section}
\title{Append section header to build log}
\usage{
.log_build_section(section_name)
}
\arguments{
\item{section_name}{Character. Name of the section.}
}
\description{
Append section header to build log
}
\keyword{internal}
