% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.log_clear_output_by_date}
\alias{.log_clear_output_by_date}
\title{Clear output logs before a specific date}
\usage{
.log_clear_output_by_date(output_dir, before_date)
}
\description{
Clear output logs before a specific date
}
\keyword{internal}
