% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.log_clear_output_by_version}
\alias{.log_clear_output_by_version}
\title{Clear output logs before a specific version}
\usage{
.log_clear_output_by_version(output_dir, before_version)
}
\description{
Clear output logs before a specific version
}
\keyword{internal}
