% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.log_dir_get_base}
\alias{.log_dir_get_base}
\title{Get the base log directory path}
\usage{
.log_dir_get_base()
}
\value{
Character. Path to the log directory.
}
\description{
Get the base log directory path
}
\keyword{internal}
