% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.log_dir_get_type}
\alias{.log_dir_get_type}
\title{Get log directory for build type}
\usage{
.log_dir_get_type(build_type = "output", create = TRUE)
}
\arguments{
\item{build_type}{Character. Either "output" or "dev".}

\item{create}{Logical. Create directory if it doesn't exist.}
}
\value{
Character. Path to the build type log directory.
}
\description{
Get log directory for build type
}
\keyword{internal}
