% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.log_file_get_most_recent}
\alias{.log_file_get_most_recent}
\title{Get the most recent log file across both output and dev builds}
\usage{
.log_file_get_most_recent()
}
\value{
Character or NULL. Path to the most recent log file, or NULL if none found.
}
\description{
Get the most recent log file across both output and dev builds
}
\keyword{internal}
