% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.log_file_get_output}
\alias{.log_file_get_output}
\title{Get output log file path for current build}
\usage{
.log_file_get_output(build_type = "output", timestamp = NULL)
}
\arguments{
\item{build_type}{Character. Either "output" or "dev".}

\item{timestamp}{Character. Timestamp in HH-MM-SS format. Default is now.}
}
\value{
Character. Path to the log file for this build.
}
\description{
Get output log file path for current build
}
\keyword{internal}
