% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{.projr_build}
\alias{.projr_build}
\title{Build project}
\usage{
.projr_build(bump_component)
}
\arguments{
\item{version}{"major", "minor", "patch".
Version to bump.
If \code{NULL}, then the lowest available
version component in the \code{version} key of \verb{_projr.yml}
is used.
Default is \code{NULL}.}
}
\description{
Build project
}
