% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{.projr_renv_rest_cleanup_lib}
\alias{.projr_renv_rest_cleanup_lib}
\title{Delete the first library path from .libPaths()}
\usage{
.projr_renv_rest_cleanup_lib()
}
\value{
NULL. Stops execution on failure.
}
\description{
Delete the first library path from .libPaths()
}
