% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{.projr_renv_rest_copy_files}
\alias{.projr_renv_rest_copy_files}
\title{Remove existing files and copy specified files into the working directory}
\usage{
.projr_renv_rest_copy_files(files_to_copy)
}
\arguments{
\item{files_to_copy}{Character vector. Paths to files to copy.}

\item{working_dir}{Character. Path to the working directory.}
}
\value{
NULL. Stops execution on failure.
}
\description{
Remove existing files and copy specified files into the working directory
}
