% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{.projr_renv_rest_disable_cache}
\alias{.projr_renv_rest_disable_cache}
\title{Disable renv cache by appending a setting to .Rprofile}
\usage{
.projr_renv_rest_disable_cache()
}
\arguments{
\item{working_dir}{Character. Path to the working directory.}
}
\value{
NULL. Stops execution on failure.
}
\description{
Disable renv cache by appending a setting to .Rprofile
}
