% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{.projr_renv_rest_init}
\alias{.projr_renv_rest_init}
\title{Initialize renv in the specified working directory using Rscript}
\usage{
.projr_renv_rest_init()
}
\arguments{
\item{working_dir}{Character. Path to the working directory.}
}
\value{
NULL. Stops execution on failure.
}
\description{
Initialize renv in the specified working directory using Rscript
}
