% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{.projr_renv_rest_restore}
\alias{.projr_renv_rest_restore}
\title{Run renv::restore() via Rscript in the specified working directory}
\usage{
.projr_renv_rest_restore()
}
\arguments{
\item{working_dir}{Character. Path to the working directory.}
}
\value{
Logical. TRUE if restore succeeds, FALSE otherwise.
}
\description{
Run renv::restore() via Rscript in the specified working directory
}
