% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{.projr_renv_rest_run_rscript}
\alias{.projr_renv_rest_run_rscript}
\title{Run an R command using Rscript in the specified working directory}
\usage{
.projr_renv_rest_run_rscript(expr, message_path = TRUE, vanilla = FALSE)
}
\arguments{
\item{expr}{Character. R expression to execute.}

\item{working_dir}{Character. Path to the working directory.}
}
\value{
List with elements 'success' (logical) and 'error' (character).
}
\description{
Run an R command using Rscript in the specified working directory
}
