% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml-dest.R
\name{.projr_yml_dest_add}
\alias{.projr_yml_dest_add}
\title{Add a remote as a destination}
\usage{
.projr_yml_dest_add(
  role,
  type,
  title,
  profile,
  content = NULL,
  structure = NULL,
  path = NULL,
  path_append_label = NULL,
  overwrite = FALSE,
  public = FALSE,
  category = NULL,
  description = NULL,
  id = NULL,
  id_parent = NULL,
  get_sync_approach = NULL,
  get_conflict = NULL,
  send_cue = NULL,
  send_sync_approach = NULL,
  send_version_source = NULL,
  send_conflict = NULL
)
}
\arguments{
\item{type}{"github", "osf" or "local".
Type of remote to create.
Selecting "github" means a GitHub release,
"osf" means an OSF node (project or component),
and "local" means a local directory.
Must be selected.}

\item{title}{character.
Title of the remote.
in which case the GitHub releases's tag
will the project version (i.e. the result of \code{projr_version_get()})
at the time of the upload.
Has no effect for \code{local} remotes.
For GitHub releases, can use title as \verb{@version},
Note that this implies that a new tag will be created with each
new version, so do not use with large files.
If not supplied, then will
automatically be generated from \code{content}.}

\item{content}{character vector.
Labels in the \code{directories} key of \code{projr_yml_get()}
to send to the remote e.g. \code{data-raw}, \code{cache}, \code{output}.}

\item{description}{character.
Description of the remote.
Has no effect for \code{local} remotes.
Default is \code{NULL}.}
}
\description{
Add a remote as a destination
}
