% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml.R
\name{.projr_yml_get}
\alias{.projr_yml_get}
\title{Get active \code{projr} settings and do no check}
\usage{
.projr_yml_get(profile)
}
\arguments{
\item{profile}{character.
If supplied, the specific profile file to read in.
"default" loads \verb{_projr.yml}, but another value
loads \verb{_projr-<profile>.yml}.
If NULL, then the active profile is used.
It not supplied, then treated as \code{NULL}.}
}
\value{
A named list.
}
\description{
A list of the active \code{projr} settings,
without doing any error checking.
Gets active \code{projr} settings, which merges settings and resolves conflicts
between local (\verb{_projr-local.yml}), profile (\verb{_projr-<projr>.yml} or as
\if{html}{\out{<key>}}-\if{html}{\out{<profile>}} keys in \verb{_projr.yml})
and default (\verb{_projr.yml}) settings.
Where there are conflicts, local settings has highest precedence
(i.e. are always preferred) and default settings have lowest precedence
(i.e. are never preferred).
}
\seealso{
projr_yml_get,projr_yml_check
}
