% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-github-httr.R
\name{.remote_rm_github_httr}
\alias{.remote_rm_github_httr}
\title{Delete a GitHub release using httr}
\usage{
.remote_rm_github_httr(repo, tag, api_url = NULL, token = NULL)
}
\arguments{
\item{repo}{Character string. Repository in format "owner/repo".}

\item{tag}{Character string. Release tag to delete.}

\item{api_url}{Optional GitHub API base URL.}

\item{token}{Optional GitHub token.}
}
\value{
Logical TRUE on success, FALSE if not found.
}
\description{
Delete a GitHub release using httr
}
\keyword{internal}
