% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml-scripts.R
\name{.yml_scripts_get_dev}
\alias{.yml_scripts_get_dev}
\title{Get list of scripts to build for dev builds}
\usage{
.yml_scripts_get_dev(profile)
}
\arguments{
\item{profile}{Profile name}
}
\value{
Vector of script paths to build or NULL
}
\description{
Checks for build.scripts.dev first, then falls back to build.scripts.build
}
\keyword{internal}
