% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml-scripts.R
\name{.yml_scripts_get_hooks_post}
\alias{.yml_scripts_get_hooks_post}
\title{Get post-build hooks from build.scripts}
\usage{
.yml_scripts_get_hooks_post(profile)
}
\arguments{
\item{profile}{Profile name}
}
\value{
Hook configuration or NULL
}
\description{
Get post-build hooks from build.scripts
}
\keyword{internal}
