% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proj_dat.R
\name{get_proj_dat}
\alias{get_proj_dat}
\title{Get project data}
\usage{
get_proj_dat(dir)
}
\arguments{
\item{dir}{A path}
}
\value{
Path and git remote url if available
}
\description{
Get project data
}
\examples{
get_proj_dat(".")
}
