% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects.R
\name{get_projects}
\alias{get_projects}
\title{Get projects}
\usage{
get_projects(dir)
}
\arguments{
\item{dir}{A path}
}
\value{
vector of project directories
}
\description{
Get projects
}
\examples{
get_projects(".")
}
