% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{prj_activate}
\alias{prj_activate}
\title{Activate project}
\usage{
prj_activate(
  wd_var = "LOCAL_WORKSPACE_FOLDER",
  path_yml = "_projr.yml",
  create_var = TRUE,
  env_var = .GlobalEnv,
  silent = FALSE
)
}
\description{
Activate project
}
