% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{projr_activate}
\alias{projr_activate}
\title{Activate project}
\usage{
projr_activate(
  wd_var = "PROJR_WORKING_DIRECTORY",
  path_yml = "_projr.yml",
  create_var = TRUE,
  env_var = .GlobalEnv,
  silent = FALSE
)
}
\arguments{
\item{silent.}{Logical.
If \code{TRUE}, then messages are suppressed
(but not warnings or errors).
Default is \code{FALSE}.}
}
\description{
Activate project
}
