% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{projr_build}
\alias{projr_build}
\title{Build project}
\usage{
projr_build(version = NULL)
}
\arguments{
\item{version}{"major", "minor", "patch", "dev" or NULL,
depending on what you chose.
Default is \code{NULL}, which}
}
\description{
Build project
}
