% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{projr_build_dev}
\alias{projr_build_dev}
\title{Build dev project}
\usage{
projr_build_dev(bump = FALSE, wd_var = "PROJR_WORKING_DIRECTORY")
}
\description{
Builds project to output, which
means recording the input and output data hashes,
building the actual bookdown document and
saving and archiving selected output.
}
