% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{projr_bump_version_dev}
\alias{projr_bump_version_dev}
\title{Bump dev version}
\usage{
projr_bump_version_dev()
}
\description{
Increments development version component
in project version by 1.
Note that this only increments the version
in \verb{_bookdown.yml} and not DESCRIPTION.
The version in DESCRIPTION will be incremented once
the project is built.
}
