% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf-yml-dest.R
\name{projr_dest_add_osf}
\alias{projr_dest_add_osf}
\alias{projr_dest_add_osf_proj}
\alias{projr_dest_add_osf_comp}
\title{Add an OSF node as a destination}
\usage{
projr_dest_add_osf(
  title,
  content = NULL,
  id = NULL,
  remote_structure = NULL,
  body = NULL,
  public = FALSE,
  category = NULL,
  parent_id = NULL,
  parent_title = NULL,
  path = NULL,
  path_append_label = NULL,
  download_sync_approach = NULL,
  download_conflict = NULL,
  upload_cue = NULL,
  upload_sync_approach = NULL,
  upload_version_source = NULL,
  upload_conflict = NULL,
  overwrite = FALSE
)

projr_dest_add_osf_proj(
  title,
  body = NULL,
  content = NULL,
  public = FALSE,
  id = NULL
)

projr_dest_add_osf_comp(
  title,
  body = NULL,
  content = NULL,
  public = FALSE,
  category = NULL,
  parent_title = NULL,
  parent_id = NULL,
  id = NULL
)
}
\arguments{
\item{title}{character. The title of the project or component.
Must be supplied.}

\item{content}{character vector. Keys to include in the OSF upload.}

\item{id}{character. The id of the project or component. Must be five
characters. Default is \code{NULL}.}

\item{body}{character. Description on OSF. Default is \code{NULL}.}

\item{public}{logical. Whether the project or component is public on OSF.
Default is \code{FALSE}.}

\item{category}{character. The category of the project or component.
The following are valid options: \code{"project"}, \code{"analysis"},
\code{"communication"}, \code{"data"}, \code{"hypothesis"}, \code{"instrumentation"},
\code{"methods and measures"}, \code{"procedure"}, \code{"project"}, \code{"software"} and other
\code{"other"}. Default is \code{NULL.}}

\item{parent_id}{character. The id of the parent project or component.
Must be five characters. Default is \code{NULL}.}

\item{parent_title}{character. The title of the parent project or component.
Default is \code{NULL}.}

\item{overwrite}{logical. If \code{TRUE}, then the project or component
will be overwitten if it exists.
Note that any components of the project/component
will be deleted from the YAML file. If \code{FALSE}, then an error will be thrown.
Default is \code{FALSE}.
element in}

\item{create}{logical. Whether to create the project or component on OSF.
Default is \code{FALSE}.}
}
\description{
Add a project or component to the _projr.yml
specification for OSF.
}
