% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{projr_dir_create}
\alias{projr_dir_create}
\title{Create a directory in _projr.yml}
\usage{
projr_dir_create(label)
}
\description{
Creates a directory that is
listed in _projr.yml for the current projr profile.
Will add to \code{.gitignore} and \code{.Rbuildignore}
as well if required.
}
