% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{projr_dir_create_all}
\alias{projr_dir_create_all}
\title{Create all directories in _projr.yml}
\usage{
projr_dir_create_all()
}
\description{
Convenience function to create all directories
listed in \code{_projr.yml} for the current projr profile.
}
\seealso{
projr_dir_get
}
