% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{projr_dir_get}
\alias{projr_dir_get}
\title{Return path to profile-specific directory}
\usage{
projr_dir_get(type, ..., rel = TRUE)
}
\arguments{
\item{type}{character.
One of \code{"data_raw"}, \code{"cache"},\code{"output"},
\code{"archive"} and \code{"bookdown"}.
Class of directory to return.
The \code{"bookdown"} option returns the path to
the output directory from \code{bookdown::render_book}
(as specified in \code{"_bookdown.yml"}),
whereas the others returns paths as specified in \code{"_projr.yml"}.}

\item{...}{Specifies sub-directory of directory returned.
Passed to \code{file.path}.}

\item{force_relative}{logical.
If \code{TRUE}, then forces that the returned
path is relative to the project root.
Default is \code{FALSE}.}
}
\value{
Character.
Path to directory requested.
}
\description{
Returns path to \code{projr} profile-specific directory.
Also creates the directory if it does not exist, and
ignores it if requested by \verb{_projr.yml}.
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
