% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{projr_dir_ignore}
\alias{projr_dir_ignore}
\title{Update ignore settings as per projr config}
\usage{
projr_dir_ignore(git_skip_adjust = NULL)
}
\arguments{
\item{git_skip_adjust}{logical.
Whether to use Git's ability to
ignore already tracked files if the
files were previously tracked but we now no
longer wish to track them.
If \code{NULL}, then treated as \code{TRUE}.
Default is \code{NULL}.}
}
\description{
Ensure \code{projr} directories are ignored
appropriately (in \code{.gitignore} and \code{.Rbuildignore})
as per \verb{_projr.yml}.

git_skip_adjust logical.
If \code{TRUE}, then the directories
are skipped by Git if already tracked
and they are supposed to be ignored
as per \code{_projr.yml}.
Default is \code{NULL}, in which
case the settings as per \code{_projr.yml}
are used (or the default, if unspecified there).
}
