% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore-manual.R
\name{projr_ignore_manual}
\alias{projr_ignore_manual}
\alias{projr_ignore_manual_dir}
\alias{projr_ignore_manual_file}
\alias{projr_ignore_manual_file_git}
\alias{projr_ignore_manual_dir_git}
\alias{projr_ignore_manual_file_rbuild}
\alias{projr_ignore_manual_dir_rbuild}
\title{Manually Ignore Files or Directories in \code{.gitignore} and \code{.Rbuildignore}}
\usage{
projr_ignore_manual(ignore)

projr_ignore_manual_dir(ignore)

projr_ignore_manual_file(ignore)

projr_ignore_manual_file_git(ignore)

projr_ignore_manual_dir_git(ignore)

projr_ignore_manual_file_rbuild(ignore)

projr_ignore_manual_dir_rbuild(ignore)
}
\arguments{
\item{ignore}{A character vector of file or directory paths to be ignored.
Paths must be valid non-empty strings.}
}
\value{
Invisibly returns \code{TRUE} if the operation succeeds, or \code{FALSE} if the input
contains invalid (empty) paths.
}
\description{
These functions allow manual addition of files and directories to the
\code{.gitignore} and \code{.Rbuildignore} files, outside of the automatic management
provided by the \code{projr} package.
\itemize{
\item .ignore_manual\verb{: General function to add both files and directories  to both }.gitignore\code{and}.Rbuildignore`. If a path does not exist, it is
treated as a file.
\item .ignore_manual_dir\verb{: Specifically adds directories to both }.gitignore\code{ and}.Rbuildignore`.
\item .ignore_manual_file\verb{: Specifically adds files to both }.gitignore\code{ and}.Rbuildignore`.
\item .ignore_manual_dir_git\verb{ and .ignore_manual_file_git}: Add
directories or files explicitly to \code{.gitignore}.
\item .ignore_manual_dir_rbuild\verb{ and .ignore_manual_file_rbuild}: Add
directories or files explicitly to \code{.Rbuildignore}.
}
}
\details{
These functions provide fine-grained control for cases where users want to
manually ignore specific paths permanently. They do not interact with the
automated ignore management system of \code{projr}.
\itemize{
\item Non-existent paths provided to .ignore_manual` are assumed to be files.
\item For \code{.gitignore}, directories are automatically appended with \verb{/**} if
missing, ensuring proper Git ignore syntax.
\item For \code{.Rbuildignore}, paths are converted to regular expressions using
\code{glob2rx} for compatibility with R's build tools.
}
}
\examples{
# Manually ignore files and directories
.ignore_manual(c("output", "tempfile.log"))

# Specifically ignore directories
.ignore_manual_dir("data")

# Specifically ignore files
.ignore_manual_file("README.md")

}
\seealso{
.ignore_auto\verb{ for dynamically managed ignore entries, and .unignore_manual}
for forcing certain paths to not be ignored.
}
