% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-helper.R
\name{projr_init_renviron}
\alias{projr_init_renviron}
\title{Set environment variables for projr_init}
\usage{
projr_init_renviron()
}
\description{
Set environment variables for \code{projr_init.}
When set, \code{projr_init} will use these variables
to provide options to populate the metadata.
This function creates the \code{.Renviron} file
for the user if it does not exist.
It then adds the variables to the \code{.Renviron} file
without any values set.
}
