% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{projr_manifest_info}
\alias{projr_manifest_info}
\title{Get Manifest Storage Information}
\usage{
projr_manifest_info()
}
\value{
A list with components:
\describe{
\item{versions}{Character vector of available versions}
\item{n_versions}{Number of versions with split manifests}
\item{split_manifest_exists}{Logical - whether split manifest directory exists}
\item{consolidated_exists}{Logical - whether consolidated manifest.csv exists}
\item{split_total_size}{Total size of split manifest files in bytes}
\item{consolidated_size}{Size of consolidated manifest.csv in bytes}
}
}
\description{
Returns information about manifest storage, including the number of versions,
file sizes, and storage format (split vs consolidated).
}
\examples{
\dontrun{
# Get manifest storage info
info <- projr_manifest_info()
print(info)
}

}
