% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-osf.R
\name{projr_osf_create_project}
\alias{projr_osf_create_project}
\title{Create a new project on OSF}
\usage{
projr_osf_create_project(title, description, public)
}
\arguments{
\item{title}{character. Title of the project.}

\item{description}{character. Description of the project.}

\item{public}{logical.
Whether the project should be public (TRUE) or private (FALSE).}
}
\value{
A character string containing the ID of the newly created project.
}
\description{
This function creates a new project on the Open Science Framework (OSF)
with the specified title, description, and visibility settings.
}
\examples{
\dontrun{
projr_osf_create_project(
  title = "My New Project",
  description = "This is a description of my new project.",
  public = TRUE # because open science
)
}

}
\seealso{
\url{https://osf.io/} for more information about OSF.
}
