% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf-config-dir.R
\name{projr_osf_source_add}
\alias{projr_osf_source_add}
\title{Add an OSF node}
\usage{
projr_osf_source_add(
  label,
  id = NULL,
  title = NULL,
  create = TRUE,
  body = NULL,
  public = FALSE,
  category = NULL,
  parent_id = NULL,
  parent_title = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{label}{character.
The directory label.
Must be supplied.}

\item{id}{character. The id of the project or component. Must be five
characters. Default is \code{NULL}, in which case a new component
is generated if \code{create} is \code{TRUE}.}

\item{title}{character.
The title of the project or component.
If not supplied, then will be set to \code{label}.}

\item{create}{logical.
Whether to create the project or component on OSF if the \code{id} is
not provided.
Default is \code{FALSE}.}

\item{body}{character vector.
Description on OSF. Default is \code{NULL}.}

\item{public}{logical.
Whether the project or component is public on OSF.
Default is \code{FALSE}.}

\item{category}{character.
The category of the project or component.
The following are valid options: \code{"project"}, \code{"analysis"},
\code{"communication"}, \code{"data"}, \code{"hypothesis"}, \code{"instrumentation"},
\code{"methods and measures"}, \code{"procedure"}, \code{"project"}, \code{"software"} and other
\code{"other"}. Default is \code{NULL.}}

\item{parent_id}{character. The id of the parent project or component.
Must be five characters (if supplied). Default is \code{NULL}.}

\item{overwrite}{logical.
If \code{TRUE}, then any pre-existing source will be overwritten
in the YAML file.
If \code{FALSE}, then a pre-existing source will throw an error.
Default is \code{TRUE}.}
}
\description{
Add a project or component to the _projr.yml
specification for OSF.
}
