% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{projr_osf_yml_up_add}
\alias{projr_osf_yml_up_add}
\alias{projr_osf_yml_up_add_proj}
\alias{projr_osf_yml_up_add_comp}
\title{Add an OSF node}
\usage{
projr_osf_yml_up_add(
  title,
  create = TRUE,
  body = NULL,
  content = NULL,
  public = FALSE,
  category = NULL,
  id = NULL,
  parent_id = NULL,
  parent_title = NULL,
  overwrite = FALSE
)

projr_osf_yml_up_add_proj(
  title,
  body = NULL,
  content = NULL,
  public = FALSE,
  id = NULL
)

projr_osf_yml_up_add_comp(
  title,
  body = NULL,
  content = NULL,
  public = FALSE,
  category = NULL,
  parent_title = NULL,
  parent_id = NULL,
  id = NULL
)
}
\arguments{
\item{title}{character. The title of the project or component.
Must be supplied.}

\item{create}{logical. Whether to create the project or component on OSF.
Default is \code{FALSE}.}

\item{body}{character. Description on OSF. Default is \code{NULL}.}

\item{content}{character vector. Keys to include in the OSF upload.}

\item{public}{logical. Whether the project or component is public on OSF.
Default is \code{FALSE}.}

\item{category}{character. The category of the project or component.
The following are valid options: \code{"project"}, \code{"analysis"},
\code{"communication"}, \code{"data"}, \code{"hypothesis"}, \code{"instrumentation"},
\code{"methods and measures"}, \code{"procedure"}, \code{"project"}, \code{"software"} and other
\code{"other"}. Default is \code{NULL.}}

\item{id}{character. The id of the project or component. Must be five
characters. Default is \code{NULL}.}

\item{parent_id}{character. The id of the parent project or component.
Must be five characters. Default is \code{NULL}.}

\item{parent_title}{character. The title of the parent project or component.
Default is \code{NULL}.}

\item{overwrite}{logical. If \code{TRUE}, then the project or component
will be overwitten as specified.
Note that any components of the project/component
will be deleted from the YAML file. Default is \code{FALSE}.}
}
\description{
Add a project or component to the _projr.yml
specification for OSF.
}
