% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{projr_path_get}
\alias{projr_path_get}
\title{Return path}
\usage{
projr_path_get(
  label,
  ...,
  create = TRUE,
  relative = FALSE,
  absolute = FALSE,
  safe = TRUE
)
}
\arguments{
\item{label}{character.
One of \code{"raw_data"}, \code{"cache"},\code{"output"},
\code{"archive"} and \code{"docs"}.
Class of directory to return.
The \code{"docs"} option returns the path to
the output directory from \code{bookdown::render_book}
(as specified in \code{"_bookdown.yml"}),
whereas the others returns paths as specified in \code{"_projr.yml"}.}

\item{...}{Specifies sub-path of directory returned.}

\item{create}{logical.
If \code{TRUE}, then the parent directory
is created if it does not exist and
it is ignored (or not) from \code{.gitignore}
and \code{.Rbuildignore} as specified
in \code{_projr.yml}.
Default is \code{TRUE}.}
}
\value{
Character.
Path to directory requested.
}
\description{

}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
