% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{projr_path_get_dir}
\alias{projr_path_get_dir}
\title{Return path to profile-specific directory}
\usage{
projr_path_get_dir(
  label,
  ...,
  create = TRUE,
  format = c("auto", "relative", "absolute"),
  safe = TRUE
)
}
\arguments{
\item{label}{character.
One of \code{"raw"}, \code{"cache"},\code{"output"},
\code{"archive"} and \code{"docs"}.
Class of directory to return.
The \code{"docs"} option returns the path to
the output directory from \code{bookdown::render_book}
(as specified in \code{"_bookdown.yml"}),
whereas the others returns paths as specified in \code{"_projr.yml"}.}

\item{...}{Specifies sub-directory of directory returned.
Passed to \code{file.path}.}

\item{create}{logical.
If \code{TRUE}, then the directory
is created if it does not exist and
it is ignored (or not) from \code{.gitignore}
and \code{.Rbuildignore} as specified
in \code{_projr.yml}.
Default is \code{TRUE}.}

\item{format}{character.
One of \code{"auto"}, \code{"relative"} or \code{"absolute"}.
If \code{"auto"}, then the format of the path returned
is inherited from the format of the path specified in \code{_projr.yml}.
If \code{"relative"}, then the path returned is relative to the current working directory.
If \code{"absolute"}, then the path returned is absolute.}

\item{safe}{logical.
If \code{TRUE}, then the output directory
is set to be \code{"<path_to_cache>.output"}
instead of \code{<path_to_output>} (as specified in \code{_projr.yml}).
The only time that this should be set to \code{TRUE}
should be when .build_output\verb{ is being run, as otherwise "development" or test runs will add to, delete or overwrite fabciles from the previous run of .build_output}.
Default is \code{TRUE}.
Do not change this unless you know what you are doing.}
}
\value{
Character.
Path to directory requested.
}
\description{
Returns path to \code{projr} profile-specific directory.
Also creates the directory if it does not exist, and
ignores it if requested by \verb{_projr.yml}.
If the path is relative, then it is calculated with respect to the project directory, but returned relative to the current working directory.
}
