% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{projr_profile_create}
\alias{projr_profile_create}
\title{Add projr profile}
\usage{
projr_profile_create(profile = NULL, silent = FALSE)
}
\arguments{
\item{profile}{character.
If not \code{NULL}, then this is the name of the profile.
If \code{NULL} and the environment variable \code{PROJR_PROFILE} is set,
then made equal to its value.}

\item{silent}{logical.
If \code{TRUE}, then any messages are suppressed.
Default is \code{TRUE} is the environment
variable \code{PROJR_TEST} is \code{"TRUE"}.}
}
\value{
Invisibly returns the new projr profile.
}
\description{
Creates a new \code{projr} profile that overrides
settings in \verb{_projr.yml}.
The profile becomes active when either the working directory
(\code{rprojroot::is_r_package$find_file()}) or
\code{Sys.getenv("PROJR_PROFILE")} is equal to \code{profile}.
}
\details{
Note that if any setting in \code{projr} profile-specific setting
is empty (indicated by \code{~} in the YAML file by default), then
the corresponding setting in \verb{_projr.yml} is used.
}
\seealso{
projr_profile_create_local,projr_yml_get
}
