% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{projr_profile_delete}
\alias{projr_profile_delete}
\title{Delete a projr profile from _projr.yml}
\usage{
projr_profile_delete(profile)
}
\arguments{
\item{profile}{character.
\code{projr} profile to delete.}
}
\value{
\code{invisible(TRUE)}.
}
\description{
Deletes a \code{projr} profile from \code{_projr.yml}
and/or its corresponding \verb{_projr-<profile>.yml} file.
}
