% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{projr_test_renv}
\alias{projr_test_renv}
\title{Test renv restore}
\usage{
projr_test_renv(file = NULL)
}
\arguments{
\item{file}{character.
Additional files to copy across to the new directory
before restoring.
May be useful, for example, to copy across an \code{.Renviron} file
to ensure that certain environment variables (such
as authentication tokens) are set.}
}
\value{
\verb{TRUE`` for successful restores and }FALSE`` for failed restores.
}
\description{
Tests renv restore without using the current project library.
Useful for testing renv restore in a clean environment
without using the cache.
}
\details{
Copies the renv.lock file and renv directory to a temporary directory,
along with .Rprofile and any files specified by the user.
Then runs renv::restore() on the temporary directory, without
using the first library used by the project to which
renv has been installing packages.
}
