% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{projr_use_data}
\alias{projr_use_data}
\title{\code{projr} drop-in replacement for usethis::use_data}
\usage{
projr_use_data(
  ...,
  internal = FALSE,
  overwrite = FALSE,
  compress = "bzip2",
  version = 2,
  ascii = FALSE,
  output_safe = TRUE
)
}
\description{
usethis::use_data always saves data to \verb{data/}, which
conflicts with the temporary directories used by \code{projr_build_dev}
and makes it difficult to restore after failed output builds.
}
