% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{projr_usr_add}
\alias{projr_usr_add}
\title{Add current working directory}
\usage{
projr_usr_add(silent = FALSE)
}
\value{
Invisibly returns the working directory.
}
\description{
Add current working directory
}
