% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{projr_version_bump}
\alias{projr_version_bump}
\title{Bump version}
\usage{
projr_version_bump(component = "dev", where = "bookdown", onto_dev = TRUE)
}
\arguments{
\item{component}{character.
Component in version to bump.
Default is \code{"dev"}.}

\item{where}{"bookdown" and/or "DESCRIPTION"/
Where to set the version.
If it includes \code{"bookdown"}, then the version is updated
in the \code{book_filename} and \code{output_dir} fields
of \code{_bookdown.yml}.
If it includes \code{"DESCRIPTION"}
Default is \code{c("bookdown")}.}

\item{onto_dev}{logical.
If \code{TRUE}, then the bumped version
will always have be a development version, i.e.
will end with the dev component specified.
Default is \code{TRUE}.}
}
\value{
Invisibly returns the new version.
}
\description{
Increase a component of the version.
By default, only increments the dev version
in _bookdown.yml (and not in DESCRIPTION as well).
This is because this function is intended primarily
to increase only the dev version in anticipation of the next build.
Non-dev versions should be bumped implicitly during runs of
\code{projr_build_output}.
}
