% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{projr_version_dev_bump}
\alias{projr_version_dev_bump}
\title{Bump development version}
\usage{
projr_version_dev_bump()
}
\value{
Invisibly returns the new version.
}
\description{
Increase development component of version.
}
