% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{projr_version_format_get}
\alias{projr_version_format_get}
\title{Get version format}
\usage{
projr_version_format_get()
}
\description{
Returns the version format,
which species the depth of semantic
versioning levels used (e.g. major or major, minor and patch)
as well as the separators between the versions (periods or dashes).
}
