% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{projr_version_format_set}
\alias{projr_version_format_set}
\title{Get version format}
\usage{
projr_version_format_set(version_format)
}
\arguments{
\item{version_format}{character.
The version format.
Specifies the version components ("major" and/or "minor" and/or "patch", and "dev"/"1"/"9000"),
as well as the separators between them.
If "1" is used instead of "dev", then "dev" component resets at "1".
If "9000" is used instead of "dev", then "dev" component resets at "9000".
If "dev" is used, then "dev" component resets at "9000".
Examples: "major.minor-1", "major-minor-patch-dev", "major.9000".
Default is "major.minor.patch.dev".}
}
\description{
Returns the version format,
which species the depth of semantic
versioning levels used (e.g. major or major, minor and patch)
as well as the separators between the versions (periods or dashes).
}
\examples{
\dontrun{
projr_version_format_set("major.minor-dev")
projr_version_format_set("major.minor.patch-1")
}
}
