% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{projr_version_set}
\alias{projr_version_set}
\title{Set project version}
\usage{
projr_version_set(version)
}
\arguments{
\item{version}{character.
Version to set.
May be dev version (i.e. include the dev component) or not.}
}
\description{
Set project version manually.
Sometimes this may be necessary if, for example, a collaborator has
built and pushed the project whilst you've made updates, and so you want
to manually increase your version
(rather than merging their changes in first).
}
