% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml.R
\name{projr_yml_check}
\alias{projr_yml_check}
\title{Check active \code{projr} settings.}
\usage{
projr_yml_check(yml_projr = NULL)
}
\arguments{
\item{path_yml}{character.
Path to YAML file.
If \code{NULL} (the default), then
checks the}
}
\value{
Returns \code{TRUE} if all checks pass.
Otherwise throws an error.
}
\description{
Checks correctness of active \code{projr} settings.
}
