% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml-dest-wrapper.R
\name{projr_yml_dest_add_local}
\alias{projr_yml_dest_add_local}
\title{Add a local directory as a destination}
\usage{
projr_yml_dest_add_local(
  title,
  content,
  path,
  structure = NULL,
  path_append_label = NULL,
  overwrite = TRUE,
  send_cue = NULL,
  send_strategy = NULL,
  send_version_source = NULL,
  send_conflict = NULL,
  profile = "default"
)
}
\arguments{
\item{title}{character.
The name of the directory.
Has no effect besides helping structure \verb{_projr.yml}.
If not supplied, will be made equal to \code{content}.}

\item{content}{character vector.
Labels of directories to include in the upload.
Options are the labels of directories
in the active \code{projr} configuration,
as well as "docs", "data" and "code".
"docs" means the directory where the documents are rendered to,
"data" means the files in the \code{"data"} directory,
and "code" means all files tracked by the Git repository.}

\item{path}{character.
Path to the directory.
If a relative path is given, then it is taken as
relative to the project home directory.
Must be supplied.}

\item{structure}{"latest" or "version".
Structure of the remote.
If "latest", then \code{path} simply contains
the latest versions of the contents.
If "version", then \code{path} will contain
a directory for each version.
If not supplied, will be \code{version}.}

\item{path_append_label}{logical.
Whether to append the label to the path.
If \code{TRUE}, then the label will be appended to \code{path}.
If \code{FALSE}, then the path will be the path to the label.
If not set, then will be treated as \code{TRUE}.}

\item{overwrite}{logical.
Whether to rewrite an existing entry of the same
title in the specified \code{projr} configuration file.
Default is TRUE.}

\item{send_cue}{TRUE/FALSE, or one of "build", "dev", "patch",
"minor" or "major".
Minimum component bumped in a project build to initiate
the upload.
If \code{TRUE}, then will be set to \code{"patch"}.
If \code{FALSE}, then will never be uploaded.
If \code{"build"}, then will be uploaded on every build,
including \code{dev} builds, so \code{dev} and \code{"build"} are equivalent.}

\item{send_strategy}{"upload-all", "upload-missing",
"sync-using-deletiong" and "sync-using-version".
How to synchronise to the remote.
If \code{upload-all}, then all files are uploaded.
If \code{upload-missing}, then only missing files are uploaded.
If \code{sync-using-deletion}, then all files on the remote
are deleted before uploading all local files.
If \code{sync-using-version}, then files
that have changed or been added locally are uploaded to the remote,
and files that have been removed locally are removed from the remote.
If not set, then "sync-using-version" will be used.}

\item{send_version_source}{"manifest" or "file".
For \code{sync-using-version} synchronisation approach,
whether to use the recorded versions of objects to
determine what has changed  ("manifest"),
or to download everything from the remote, version it
and compare it to what's in the local folder ("file").
If not set, then "manifest" is used.}

\item{send_conflict}{"overwrite", "error" or "skip".
What to do if a file that is to be uploaded
to the remote is already on the remote.
Default is "overwrite".}

\item{profile}{character.
Profile to write the settings to.
If "default", then written to \verb{_projr.yml},
otherwise written to \verb{_projr-<profile>.yml}.
The default is "default".}
}
\description{
Add a local directory as a destination
to a _projr.yml file.
}
