% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml.R
\name{projr_yml_get}
\alias{projr_yml_get}
\title{Get settings for active profile profile}
\usage{
projr_yml_get()
}
\value{
A named list.
}
\description{
Reads in \code{_projr.yml} and returns
it after setting the directories to the directories
of the currently active profile.
}
