% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml-cite.R
\name{projr_yml_cite_set}
\alias{projr_yml_cite_set}
\title{Set citation options}
\usage{
projr_yml_cite_set(
  all_three = NULL,
  codemeta = NULL,
  cff = NULL,
  inst_citation = NULL,
  profile = "default"
)
}
\arguments{
\item{all_three}{logical.
Whether to generate all three files.
If \code{NULL}, then \code{codemeta}, \code{cff} and \code{inst_citation} are used.}

\item{codemeta}{logical.
Whether to generate a codemeta.json file.
If \code{NULL}, then setting is not changed.}

\item{cff}{logical.
Whether to generate a CITATION.cff file.
If \code{NULL}, then setting is not changed.}

\item{inst_citation}{logical.
Whether to generate a CITATION file in the inst/ directory.
If \code{NULL}, then setting is not changed.}

\item{profile}{character.
Profile to add the script to.
If \verb{"default"`` (the default), the script is added to the default profile, which is }_projr.yml`.}
}
\description{
Set citation options for the project.
The options are:
\itemize{
\item{codemeta}{Whether to generate a codemeta.json file.}
\item{cff}{Whether to generate a CITATION.cff file.}
\item{inst-citation}{Whether to generate a CITATION file in the inst/ directory.}
}
The default is to leave all the settings unchanged.
}
