% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-output.R
\name{.cli_debug}
\alias{.cli_debug}
\title{Show a debug message (only shown at debug level)}
\usage{
.cli_debug(..., .envir = parent.frame())
}
\arguments{
\item{...}{Message components passed to cli::cli_text}

\item{.envir}{Environment for variable evaluation}
}
\description{
Show a debug message (only shown at debug level)
}
\keyword{internal}
