% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-output.R
\name{.cli_message}
\alias{.cli_message}
\title{Show a standard informational message}
\usage{
.cli_message(
  ...,
  log_type,
  console_level = "std",
  console_fn,
  .envir = parent.frame()
)
}
\arguments{
\item{...}{Message components passed to cli::cli_alert_info}

\item{log_type}{Character. Type of log message ("info", "debug", etc.).}

\item{console_level}{Character. Minimum output level to show in console.}

\item{console_fn}{Function. CLI function to use for console output.}

\item{.envir}{Environment for variable evaluation}
}
\description{
Show a standard informational message
}
\keyword{internal}
