% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-output.R
\name{.cli_step}
\alias{.cli_step}
\title{Show a step message (sub-item in a stage)}
\usage{
.cli_step(..., output_level = "std", .envir = parent.frame(), log_file = NULL)
}
\arguments{
\item{...}{Message components passed to cli::cli_li}

\item{output_level}{Character. Current output level.}

\item{.envir}{Environment for variable evaluation}

\item{log_file}{Character. Path to log file (optional).}
}
\description{
Show a step message (sub-item in a stage)
}
\keyword{internal}
