% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.log_history_add}
\alias{.log_history_add}
\title{Update build history file}
\usage{
.log_history_add(
  build_type = "output",
  bump_component = NULL,
  msg = "",
  success = TRUE
)
}
\arguments{
\item{build_type}{Character. Either "output" or "dev".}

\item{bump_component}{Character. Version bump component.}

\item{msg}{Character. Build message.}

\item{success}{Logical. Whether build succeeded.}
}
\description{
Update build history file
}
\keyword{internal}
