% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml-scripts.R
\name{.yml_scripts_get_build}
\alias{.yml_scripts_get_build}
\title{Get list of scripts to build for production builds}
\usage{
.yml_scripts_get_build(profile)
}
\arguments{
\item{profile}{Profile name}
}
\value{
Vector of script paths to build or NULL
}
\description{
Supports two formats:
\enumerate{
\item Direct list: scripts: \link{file1.qmd, file2.qmd}
\item Under build key: scripts: { build: \link{file1.qmd, file2.qmd} }
}
}
\keyword{internal}
