% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{projr_instr_auth_github}
\alias{projr_instr_auth_github}
\title{Two-minutes or less authorisation instructions}
\usage{
projr_instr_auth_github()
}
\description{
Print easy-to-follow,
step-by-step instructions for authorisation
to GitHub.
}
