% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{projr_build_dev}
\alias{projr_build_dev}
\title{Build dev project}
\usage{
projr_build_dev(file = NULL, bump = FALSE, ...)
}
\arguments{
\item{file}{character vector.
Paths to files to build.
Paths may be relative to project root, or absolute.
Default is NULL, in which case all files are built.}

\item{bump}{logical.
Whether to increment dev version for build.
Default is \code{FALSE}.}

\item{...}{Arguments passed to \code{bookdown::render}.}
}
\description{
Builds project to output, which
means recording the input and output data hashes,
building the actual bookdown document and
saving and archiving selected output.
}
