% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{projr_build_output}
\alias{projr_build_output}
\title{Build project to output}
\usage{
projr_build_output(bump_component, msg = NULL, ...)
}
\arguments{
\item{bump_component}{"major", "minor", "patch" or missing.
Specifies version component to increment.
If missing, then is set equal to the lowest version component
in used version format.
No default (i.e. is missing by default).}

\item{msg}{character.
Message to append to Git commit messages.
Default is \code{NULL}, in which
case the user is prompted for a message or,
if the session is not interactive, it is
left empty.
Default is \code{NULL}.
Note that the Git messages in this case would not be blank -
they would simply consist of details as to the version
being bumped to and the stage in the build process
at which the commit was made.}

\item{...}{Arguments passed to \code{bookdown::render}.}
}
\description{
Builds project to output, which
means recording the input and output data hashes,
building the actual bookdown document and
saving and archiving selected output.
}
