% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{projr_build_output}
\alias{projr_build_output}
\title{Build project to output}
\usage{
projr_build_output(bump_component, ...)
}
\arguments{
\item{...}{Arguments passed to \code{bookdown::render}.}
}
\description{
Builds project to output, which
means recording the input and output data hashes,
building the actual bookdown document and
saving and archiving selected output.
}
