% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{projr_dir_create}
\alias{projr_dir_create}
\title{Create a directory in _projr.yml}
\usage{
projr_dir_create(label)
}
\arguments{
\item{label}{character.
One of \code{"data_raw"}, \code{"cache"},\code{"output"},
\code{"archive"} and \code{"docs"}.
Class of directory to return.
The \code{"docs"} option returns the path to
the output directory from \code{bookdown::render_book}
(as specified in \code{"_bookdown.yml"}),
whereas the others returns paths as specified in \code{"_projr.yml"}.}
}
\description{
Creates a directory that is
listed in _projr.yml for the current projr profile.
Will add to \code{.gitignore} and \code{.Rbuildignore}
as well if required.
}
