% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{projr_dir_create}
\alias{projr_dir_create}
\title{Create a directory in _projr.yml}
\usage{
projr_dir_create(label, ..., safe = TRUE)
}
\arguments{
\item{label}{character.
One of \code{"data_raw"}, \code{"cache"},\code{"output"},
\code{"archive"} and \code{"docs"}.
Class of directory to return.
The \code{"docs"} option returns the path to
the output directory from \code{bookdown::render_book}
(as specified in \code{"_bookdown.yml"}),
whereas the others returns paths as specified in \code{"_projr.yml"}.}

\item{...}{Specifies sub-directory of directory returned.
Passed to \code{file.path}.}

\item{safe}{logical.
If \code{TRUE}, then the output directory
is set to be \code{"<path_to_cache>/projr_output"}
instead of \code{<path_to_output>} (as specified in \code{_projr.yml}).
The only time that this should be set to \code{TRUE}
should be when \code{projr_build_output} is being run, as otherwise
"development" or test runs will add to, delete or overwrite fabciles
from the previous run of \code{projr_build_output}.
Default is \code{TRUE}.
Do not change this unless you know what you are doing.}
}
\description{
Creates a directory that is
listed in _projr.yml for the current projr profile.
Will add to \code{.gitignore} and \code{.Rbuildignore}
as well if required.
}
