% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{projr_dir_get}
\alias{projr_dir_get}
\title{Return path to profile-specific directory}
\usage{
projr_dir_get(type, ...)
}
\arguments{
\item{type}{character.
Class of directory to return.}

\item{...}{Specifies sub-directory of directory returned.
Passed to \code{file.path}.}
}
\value{
Character.
Path to directory requested.
}
\description{
Returns path to profile-specific directory.
Also creates the directory if it does not exist, and
ignores it if requested by \verb{_projr.yml}.
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
