% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{projr_env_file_activate}
\alias{projr_env_file_activate}
\title{Activate environment variables in files}
\usage{
projr_env_file_activate(file = NULL)
}
\arguments{
\item{file}{character vector.
Paths to files to activate.
Default is NULL, in which case all relevant files that exist are used
(\code{environment.local}, \verb{_environment-<profile>} and \verb{_environment},
in order of decreasing priority and where \verb{<profile>} is
an activated Quarto profile)}
}
\description{
Which file(s) are used to supply
default values for unspecified environment variables.
}
