% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-pre-env.R
\name{projr_env_file_activate}
\alias{projr_env_file_activate}
\title{Activate environment variables in files}
\usage{
projr_env_file_activate(file = NULL, env = NULL)
}
\description{
Activate environment variables in files
}
