% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore.R
\name{projr_ignore}
\alias{projr_ignore}
\title{Update \code{.gitignore} and \code{.Rbuildignore} with projr-managed ignores}
\usage{
projr_ignore()
}
\value{
Called primarily for its side effects (modifying
\code{.gitignore} and/or \code{.Rbuildignore}). Returns \code{TRUE} invisibly.
}
\description{
The \code{projr_ignore()} function updates the project’s \code{.gitignore} and
\code{.Rbuildignore} files to reflect directories and files managed by
projr, as well as
other directories and files that should
clearly be ignored.
}
\examples{
\dontrun{
  projr_ignore()
}

}
