% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{projr_init}
\alias{projr_init}
\title{Initialise project}
\usage{
projr_init(
  yml_path_from = NULL,
  renv_force = FALSE,
  renv_bioconductor = TRUE,
  public = FALSE
)
}
\arguments{
\item{yml_path_from}{character.
Path to YAML file to use as \verb{_projr.yml}.
If not supplied, then default \verb{_projr.yml}
file is used.}

\item{renv_force}{Logical.
Passed to \code{renv::init()}.
If \code{FALSE}, then \code{renv::init()} will not run
if it detects that the working directory
already is registered with renv.
Default is \code{FALSE}.}

\item{renv_bioconductor}{Logical.
Whether \code{renv} should look for packages
on Bioconductor.
Default is \code{TRUE}.}

\item{public}{logical.
Whether the GitHub repo created (if any)
is public or not.
Default is \code{FALSE}.}
}
\description{
Initialise project
}
\seealso{
projr_init_renviron
}
