% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{projr_par_get}
\alias{projr_par_get}
\alias{projr_param_get}
\title{Get project parameters}
\usage{
projr_par_get(..., profile = NULL)

projr_param_get(..., profile = NULL)
}
\arguments{
\item{...}{character.
Sequential names to specify path in list.
For example, .param_get("a", "b")\verb{returns the value of}projr$param$a$b`.}

\item{profile}{character.
If \code{NULL}, then the active profile is used.
Default is \code{NULL}.}
}
\description{
Get project parameters from \code{param} key
in \code{projr} configuration.
}
