% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{projr_path_get}
\alias{projr_path_get}
\title{Return path}
\usage{
projr_path_get(
  label,
  ...,
  create = TRUE,
  path_relative_force = FALSE,
  output_safe = TRUE
)
}
\arguments{
\item{label}{character.
One of \code{"data_raw"}, \code{"cache"},\code{"output"},
\code{"archive"} and \code{"docs"}.
Class of directory to return.
The \code{"docs"} option returns the path to
the output directory from \code{bookdown::render_book}
(as specified in \code{"_bookdown.yml"}),
whereas the others returns paths as specified in \code{"_projr.yml"}.}

\item{...}{Specifies sub-path of directory returned.}

\item{create}{logical.
If \code{TRUE}, then the parent directory
is created if it does not exist and
it is ignored (or not) from \code{.gitignore}
and \code{.Rbuildignore} as specified
in \code{_projr.yml}.
Default is \code{TRUE}.}

\item{path_relative_force}{logical.
If \code{TRUE}, then forces that the returned
path is relative to the project root.
Default is \code{FALSE}.}

\item{output_safe}{logical.
If \code{TRUE}, then the output directory
is set to be \code{"<path_to_cache>/projr_output"}
instead of \code{<path_to_output>} (as specified in \code{_projr.yml}).
The only time that this should be set to \code{TRUE}
should be when \code{projr_build_output} is being run, as otherwise
"development" or test runs will add to, delete or overwrite files
from the previous run of \code{projr_build_output}.
Default is \code{TRUE}.}
}
\value{
Character.
Path to directory requested.
}
\description{
Returns path to \code{projr} profile-specific directory.
Differs from \code{projr_dir_get} in that it does not assume
that the path is to a directory.

Will create the parent directory of the specified
path if it does not exist, and ignore it
if requested by \verb{_projr.yml}.
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
