% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{projr_profile_create}
\alias{projr_profile_create}
\title{Add projr profile}
\usage{
projr_profile_create()
}
\value{
Invisibly returns the new projr profile.
}
\description{
Adds either the environment variable "PROJR_PROFILE"
(if set) or the current project directory (i.e.
where the DESCRIPTION file is)
as a new projr profile.
}
