% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{projr_profile_create}
\alias{projr_profile_create}
\title{Add current working directory}
\usage{
projr_profile_create(silent = FALSE)
}
\value{
Invisibly returns the working directory.
}
\description{
Add current working directory
}
