% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{projr_profile_get}
\alias{projr_profile_get}
\title{Get active projr profile}
\usage{
projr_profile_get()
}
\value{
Character vector of length 1
corresponding to active \code{projr} profile.
}
\description{
Get active \code{projr} profile.
First looks for \verb{\code{Sys.getenv("PROJR_PROFILE")}}
and then, if that is empty, takes the current
project directory
(as returned by \code{rprojroot::is_r_package()$find_file())},
i.e. the folder containing the file \code{DESCRIPTION}.
If there is no entry in \code{_projr.yml} matching the above selected
profile, then the profile \code{"default"} is returned.
}
