% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{projr_restore}
\alias{projr_restore}
\title{Restore project artefact directory}
\usage{
projr_restore(label = NULL, type = NULL, title = NULL)
}
\arguments{
\item{label}{character vector.
Must all be labels found in \verb{_projr.yml} (or
default labels).
If \code{NULL} (the default), then all \code{raw} labels
are restored (e.g. \code{raw-data}, \code{raw-docs}, \code{raw}, etc.).}

\item{type}{character.
Type of remote to restore from.
One of \code{local}, \code{osf} and \code{github}.
If \code{NULL}, then an appropriate remote is searched
for.
This is the first destination that has
this \code{label} in its \code{source} field (or
has its \code{source} field as \code{TRUE}).
If none found, then the first type
and title combination that has this
label in its contents is used.}

\item{title}{character.
Title of remote.
Ignored if \code{type} is not specified.}
}
\value{
Invisible returns \code{TRUE} if successful.
}
\description{
Restore project artefact directory
}
\examples{
\dontrun{
projr_restore("raw-data")
}
}
