% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{projr_version_set}
\alias{projr_version_set}
\title{Set project version}
\usage{
projr_version_set(version, where = c("bookdown", "DESCRIPTION"))
}
\arguments{
\item{where}{"bookdown" and/or "DESCRIPTION"/
Where to set the version.
If it includes \code{"bookdown"}, then the version is updated
in the \code{book_filename} and \code{output_dir} fields
of \code{_bookdown.yml}.
If it includes \code{"DESCRIPTION"}
Default is \code{c("bookdown")}.}
}
\description{
Set project version manually.
Sometimes this may be necessary if, for example, a collaborator has
built and pushed the project whilst you've made updates, and so you want
to manually increase your version
(rather than merging their changes in first).
}
