% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{projr_version_set}
\alias{projr_version_set}
\title{Set project version}
\usage{
projr_version_set(version, where = c("docs", "DESCRIPTION"))
}
\arguments{
\item{version}{character.
Version to set.
May be dev version (i.e. include the dev component) or not.}

\item{where}{"docs" and/or "DESCRIPTION"/
Where to set the version.
If it includes \code{"docs"}, then the version is updated
in the \code{book_filename} and \code{output_dir} fields
of \"docs"{_bookdown.yml}.
If it includes \"docs"{"DESCRIPTION"}
Default is \code{c("docs")}.}
}
\description{
Set project version manually.
Sometimes this may be necessary if, for example, a collaborator has
built and pushed the project whilst you've made updates, and so you want
to manually increase your version
(rather than merging their changes in first).
}
