% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml.R
\name{projr_yml_check}
\alias{projr_yml_check}
\title{Check active \code{projr} settings.}
\usage{
projr_yml_check(yml_projr = NULL)
}
\arguments{
\item{yml_projr}{list.
Projr settings. If not supplied,
then the results of \code{projr_yml_get_unchecked} is used.}
}
\value{
Returns \code{TRUE} if all checks pass.
Otherwise throws an error.
}
\description{
Checks correctness of active \code{projr} settings.
}
