% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml-dest-wrapper.R
\name{projr_yml_dest_add_local}
\alias{projr_yml_dest_add_local}
\title{Add a local directory as a destination}
\usage{
projr_yml_dest_add_local(
  title,
  content,
  path,
  structure = NULL,
  overwrite = TRUE,
  send_cue = NULL,
  send_strategy = NULL,
  send_inspect = NULL,
  profile = "default"
)
}
\arguments{
\item{title}{character.
The name of the directory.
Has no effect besides helping structure \verb{_projr.yml}.
If not supplied, will be made equal to \code{content}.}

\item{content}{character vector.
Labels of directories to include in the upload.
Options are the labels of directories
in the active \code{projr} configuration.}

\item{path}{character.
Path to the directory.
If a relative path is given, then it is taken as
relative to the project home directory.
Must be supplied.}

\item{structure}{"latest" or "archive".
Structure of the remote.
If "latest", then \code{path} simply contains
the latest versions of the contents.
If "version", then \code{path} will contain
a directory for each version.
If not supplied, will be \code{archive}.}

\item{overwrite}{logical.
Whether to rewrite an existing entry of the same
title in the specified \code{projr} configuration file.
Default is TRUE.}

\item{send_cue}{"always", "if-change" or "never".
When to send to the remote. Default is NULL.}

\item{send_strategy}{"upload-all", "upload-missing",
"sync-purge" or "sync-diff".
How to synchronise to the remote. Default is NULL.}

\item{send_inspect}{"manifest" , "file" or "none".
What to look at to find what are the
files on the remote, and their versions. Default is NULL.}

\item{profile}{character.
Profile to write the settings to.
If "default", then written to \verb{_projr.yml},
otherwise written to \verb{_projr-<profile>.yml}.
The default is "default".}
}
\description{
Add a local directory as a destination
to a _projr.yml file.
}
