% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml-dest-wrapper.R
\name{projr_yml_dest_add_osf}
\alias{projr_yml_dest_add_osf}
\alias{projr_yml_dest_add_local}
\title{Add an OSF node as a destination}
\usage{
projr_yml_dest_add_osf(
  title = NULL,
  content = NULL,
  structure = NULL,
  path = NULL,
  path_append_label = NULL,
  overwrite = FALSE,
  public = FALSE,
  category = NULL,
  description = NULL,
  id = NULL,
  id_parent = NULL,
  title_parent = NULL,
  get_sync_approach = NULL,
  get_conflict = NULL,
  send_cue = NULL,
  send_sync_approach = NULL,
  send_version_source = NULL,
  send_conflict = NULL
)

projr_yml_dest_add_local(
  title = NULL,
  content = NULL,
  path = NULL,
  path_append_label = NULL,
  structure = NULL,
  get_sync_approach = NULL,
  get_conflict = NULL,
  send_cue = NULL,
  send_sync_approach = NULL,
  send_version_source = NULL,
  send_conflict = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{title}{character.
The name of the directory.
Has no effect besides helping structure \verb{_projr.yml}.
If not supplied, will be made equal to \code{content}.}

\item{content}{character vector.
Labels to include in the OSF upload, e.g. \code{data-raw}, \code{cache}, \code{output},}

\item{structure}{"latest" or "version".
Structure of the remote.
If "latest", then \code{path} simply contains
the latest versions of the contents.
If "version", then \code{path} will contain
a directory for each version.
If not supplied, will be \code{version}.}

\item{path}{character.
Path to the directory on the local system.
If a relative path is given, then it is taken as
relative to the project home directory.
If not supplied, then will be by default \verb{_output}
if structure is \code{latest} and \verb{_archive} if structure is
\code{version}.}

\item{path_append_label}{logical.
Whether to append the label to the path.
If \code{TRUE}, then the label will be appended to \code{path}.
If \code{FALSE}, then the path will be the path to the label.}

\item{overwrite}{logical. If \code{TRUE}, then the project or component
will be overwitten if it exists.
Note that any components of the project/component
will be deleted from the YAML file. If \code{FALSE}, then an error will be thrown.
Default is \code{FALSE}.
element in}

\item{public}{logical. Whether the project or component is public on OSF.
Default is \code{FALSE}.}

\item{category}{character. The category of the project or component.
The following are valid options: \code{"project"}, \code{"analysis"},
\code{"communication"}, \code{"data"}, \code{"hypothesis"}, \code{"instrumentation"},
\code{"methods and measures"}, \code{"procedure"}, \code{"project"}, \code{"software"} and other
\code{"other"}. Default is \code{NULL.}}

\item{id}{character. The id of the project or component. Must be five
characters. Default is \code{NULL}.}

\item{id_parent}{character. The id of the parent project or component.
Must be five characters. Default is \code{NULL}.}

\item{title_parent}{character. The title of the parent project or component.
Default is \code{NULL}.}

\item{body}{character. Description on OSF. Default is \code{NULL}.}

\item{ignore_git}{logical.
Whether to exclude the final path from Git.
If not supplied, then the path will be excluded.}

\item{ignore_rbuild}{logical.
Whether to exclude the final path in \code{.Rbuildignore}.
If not supplied, then the path will be excluded.}

\item{cue}{character.
When to cue the upload.}
}
\description{
Add a project or component to the _projr.yml
specification for OSF.

Add a project or component to the _projr.yml
specification for OSF.

Note: this will always add to \verb{_projr.yml}.
If you want to add to another \code{projr} file, you can either
create that manually or copy the result of applying
this function to \verb{_projr.yml} across.
}
