% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml.R
\name{projr_yml_get}
\alias{projr_yml_get}
\title{Get active \code{projr} settings and checks for validity}
\usage{
projr_yml_get(profile = NULL, check = FALSE)
}
\arguments{
\item{profile}{character.
\code{projr} profile to use.
If \code{NULL}, then the active profile is used.
Default is \code{NULL}.}

\item{check}{logical.
Whether to check the validity of the settings.
Default is \code{FALSE}.}
}
\value{
A named list, if the settings are valid.
}
\description{
Gets active \code{projr} settings, which merges settings and resolves conflicts
between local (\verb{_projr-local.yml}), profile (\verb{_projr-<projr>.yml} or as
\if{html}{\out{<key>}}-\if{html}{\out{<profile>}} keys in \verb{_projr.yml})
and default (\verb{_projr.yml}) settings.
Where there are conflicts, local settings has highest precedence
(i.e. are always preferred) and default settings have lowest precedence
(i.e. are never preferred).

Note that an error is thrown if the active settings
are invalid.
}
\seealso{
projr_yml_get_unchecked,projr_yml_check
}
