% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml-cite.R
\name{projr_yml_cite_set}
\alias{projr_yml_cite_set}
\alias{projr_yml_cite_set_default}
\title{Set citation options}
\usage{
projr_yml_cite_set(
  all = NULL,
  codemeta = NULL,
  cff = NULL,
  inst_citation = NULL,
  simplify_identical = TRUE,
  simplify_default = TRUE,
  profile = "default"
)

projr_yml_cite_set_default(
  profile = "default",
  simplify_identical = TRUE,
  simplify_default = TRUE
)
}
\arguments{
\item{all}{logical.
Whether to set all the options
to \code{TRUE} or \code{FALSE}.
If \code{NULL}, then \code{codemeta}, \code{cff} and \code{inst_citation} are used.
Default is \code{NULL}.}

\item{codemeta}{logical.
Whether to generate a codemeta.json file.
If \code{NULL}, then setting is not changed.}

\item{cff}{logical.
Whether to generate a CITATION.cff file.
If \code{NULL}, then setting is not changed.}

\item{inst_citation}{logical.
Whether to generate a CITATION file in the inst/ directory.
If \code{NULL}, then setting is not changed.}

\item{simplify_identical}{logical.
If \code{TRUE}, then if all the settings are the same
then only \code{cite: TRUE} or \code{cite: FALSE} is written to \verb{_projr.yml}.
Default is \code{TRUE}.}

\item{simplify_default}{logical.
If \code{TRUE}, then if the settings are the same as the default
(which is TRUE),
then the settings are removed from \verb{_projr.yml}.
Default is \code{TRUE}.}

\item{profile}{character.
Profile to add the script to.
If \verb{"default"`` (the default), the script is added to the default profile, which is }_projr.yml\verb{. If }NULL\verb{, then the active profile is used (i.e the merge of }_projr-local.yml\verb{, }_projr-\if{html}{\out{<profile>}}.yml\code{and}_projr.yml\verb{) and written to }_projr.yml\verb{. If another character vector, then the corresponding profile is used and written to }_projr-\if{html}{\out{<profile>}}.yml`.}
}
\description{
\code{projr_yml_cite_set} sets the citation options in \verb{_projr.yml}.

The options are:
\itemize{
\item{codemeta: }{whether to generate a codemeta.json file.}
\item{cff: }{
whether to generate a CITATION.cff file.
Indexable by GitHub and Zenodo.
}
\item{inst-citation: }{
whether to generate a CITATION file in the inst/ directory.
If the project is installed as an \code{R} package,
then citation data can be generated by
\verb{citation(package = <project_name>)}.
}
}
The default is to leave all the settings unchanged.

If these settings are not setting in \verb{_projr.yml},
then the default is to generate a codemeta.json file,
a CITATION.cff file and a CITATION file in the inst/ directory.

\code{projr_yml_cite_set_default} sets all citation options to default (\code{TRUE}).
}
\examples{
\dontrun{
# set all to TRUE
projr_yml_cite_set(all = TRUE)

# set all to FALSE
projr_yml_cite_set(all = FALSE)

# set only cff to FALSE
projr_yml_cite_set(cff = FALSE)

# revert to defaults
projr_yml_cite_set_default()
}

}
